refactor: Reflection 문제로 인해 최상위 User 객체에는 필요한 username과 password만 가지도록 변경
