+++ src/main/java/nextstep/auth/authentication/User.java
@@ -2,29 +2,21 @@ package nextstep.auth.authentication;
 
 import nextstep.member.domain.Member;
 
-import java.util.Objects;
-
 public class User {
-    private Long id;
-    private String email;
+    private String username;
     private String password;
 
-    public User(Long id, String email, String password) {
-        this.id = id;
-        this.email = email;
+    public User(String username, String password) {
+        this.username = username;
         this.password = password;
     }
 
     public static User ofMember(Member member) {
-        return new User(member.getId(), member.getEmail(), member.getPassword());
+        return new User(member.getEmail(), member.getPassword());
     }
 
-    public Long getId() {
-        return id;
-    }
-
-    public String getEmail() {
-        return email;
+    public String getUsername() {
+        return username;
     }
 
     public String getPassword() {
@@ -33,22 +25,5 @@ public class User {
 
     public boolean checkPassword(String password) {
         return this.password.equals(password);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        User that = (User) o;
-        return Objects.equals(id, that.id) && Objects.equals(email, that.email) && Objects.equals(password, that.password);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, email, password);
     }
 }
