+++ src/main/java/nextstep/member/domain/LoginMember.java
@@ -3,7 +3,12 @@ package nextstep.member.domain;
 
 import nextstep.auth.authentication.User;
 
+import java.util.Objects;
+
 public class LoginMember extends User {
+    private Long id;
+    private String email;
+    private String password;
     private Integer age;
 
     public static LoginMember of(Member member) {
@@ -11,11 +16,40 @@ public class LoginMember extends User {
     }
 
     public LoginMember(Long id, String email, String password, Integer age) {
-        super(id, email, password);
+        super(email, password);
+        this.id = id;
+        this.email = email;
+        this.password = password;
         this.age = age;
+    }
+
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getEmail() {
+        return email;
     }
 
     public Integer getAge() {
         return age;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        LoginMember that = (LoginMember) o;
+        return Objects.equals(id, that.id) && Objects.equals(email, that.email) && Objects.equals(password, that.password) && Objects.equals(age, that.age);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, email, password, age);
     }
 }
