+++ src/test/java/nextstep/auth/unit/TokenAuthenticationInterceptorTest.java
@@ -10,6 +10,7 @@ import nextstep.auth.token.JwtTokenProvider;
 import nextstep.auth.token.TokenRequest;
 import nextstep.auth.token.TokenResponse;
 import nextstep.member.application.CustomUserDetailsService;
+import nextstep.member.domain.LoginMember;
 import nextstep.support.MockRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -53,7 +54,7 @@ class TokenAuthenticationInterceptorTest {
 
     @Test
     void preHandle() throws Exception {
-        when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(new User(1L, EMAIL, PASSWORD));
+        when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(new LoginMember(1L, EMAIL, PASSWORD, 30));
         when(jwtTokenProvider.createToken(anyString())).thenReturn(JWT_TOKEN);
 
         interceptor.preHandle(request, response, new Object());
@@ -64,7 +65,7 @@ class TokenAuthenticationInterceptorTest {
     @Test
     @DisplayName("비밀번호가 일치하지 않으면 AuthenticationException 발생")
     void authenticateInvalidPassword() {
-        when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(new User(1L, EMAIL, "NEW_PASSWORD"));
+        when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(new LoginMember(1L, EMAIL, "NEW_PASSWORD", 30));
 
         사용자_검증에_실패하면_에러_발생(interceptor, request, response, new Object());
     }
